# add fusion for forward Narrow