This removes the `copy` associated with an nn.Narrow in the forward direction.

This reduces [char-rnn](http://github.com/karpathy/char-rnn) iteration time on an NVIDIA 940M and 5200U proc:
- cutorch: 0.13s => 0.12s
- cltorch: 0.27s => 0.26s
- cpu: 3.5s => 3.0s

Notes:
- It is true that in the specific this particular optimization could be done by modifying the nn.Narrow to have an `inplace` option.  However, I'd like to add a backwards fusion in the future somehow, which almost certainly could not be done inside the nn.Narrow module, could only be done using knowledge of upstream modules at nngraph level.
- putting this inside gModule.lua is almost certainly the wrong place to put this.  But in the generic, these optimizations cannot be done inside the nn modules, since they dont have knowledge of their up/downstream module
- arguably I could make this specific change for now by adding an `inplace` option to nn.Narrow, and then discuss backwards fusion at such a time as it exists?