@koraykv 2015-07-25
I would prefer going with the 3rd option right now. Doing optimizations over the graph is something that we know has to be done at some point, but I think putting module specific code into gModule is not the best way to do it. We could also start writing an optimization lib inside(or outside) nngraph that could implement many reductions. But right now, can you put inplace option into narrow? Also, the speed-ups would be generic and anyone using nn.Narrow even outside nngraph would see the benefits.