@hughperkins 2015-07-25
Ok, let's discuss once backwards shortcircuiting/fusion is working, because backwards cannot be implemented inside nn.